fixing bug #1240 Controls on contextual menu are done on client side and server side. Moreover, the contextual menu no longer appears for non-admin users if rights on topics are not used
git-svn-id: https://www.silverpeas.org/svn/silverpeas/components/trunk@2684 a8e77078-a1c7-4fa5-b8fc-53c5178a176c

+++ kmelia/kmelia-war/src/main/java/com/stratelia/webactiv/kmelia/control/KmeliaSessionController.java
@@ -886,17 +886,21 @@ public class KmeliaSessionController extends AbstractComponentSessionController 
   }
 
   public synchronized void deleteTopic(String topicId) throws RemoteException {
-    // First, remove rights on topic and its descendants
-    List<NodeDetail> treeview = getNodeBm().getSubTree(getNodePK(topicId));
-    NodeDetail node = null;
-    for (int n = 0; n < treeview.size(); n++) {
-      node = treeview.get(n);
+    NodeDetail node = getNodeHeader(topicId);
+    // check if user is allowed to delete this topic
+    if (SilverpeasRole.admin.isInRole(getUserTopicProfile(topicId)) ||
+        SilverpeasRole.admin.isInRole(getUserTopicProfile(node.getFatherPK().getId()))) {
+      // First, remove rights on topic and its descendants
+      List<NodeDetail> treeview = getNodeBm().getSubTree(getNodePK(topicId));
+      for (int n = 0; n < treeview.size(); n++) {
+        node = treeview.get(n);
 
-      deleteTopicRoles(node);
+        deleteTopicRoles(node);
+      }
+
+      // Then, remove the topic itself
+      getKmeliaBm().deleteTopic(getNodePK(topicId));
     }
-
-    // Then, remove the topic itself
-    getKmeliaBm().deleteTopic(getNodePK(topicId));
   }
 
   public synchronized void changeSubTopicsOrder(String way, String subTopicId)
